+++ chain/chain/src/lib.rs
@@ -26,7 +26,6 @@ mod chain_update;
 pub mod crypto_hash_timer;
 mod doomslug;
 pub mod flat_storage_init;
-pub mod flat_storage_resharder;
 mod garbage_collection;
 pub mod genesis;
 mod lightclient;
