+++ chain/chain/src/resharding/mod.rs
@@ -1,4 +1,5 @@
 pub mod event_type;
+pub mod flat_storage_resharder;
 pub mod manager;
 pub mod resharding_actor;
 pub mod types;
