+++ tools/flat-storage/src/resume_resharding.rs
@@ -1,5 +1,5 @@
 use crate::commands::ResumeReshardingCmd;
-use near_chain::flat_storage_resharder::FlatStorageResharder;
+use near_chain::resharding::flat_storage_resharder::FlatStorageResharder;
 use near_chain_configs::ReshardingHandle;
 use near_epoch_manager::EpochManager;
 use near_store::{ShardUId, StoreOpener};
